+++ gameObjects/Ghost.js
@@ -4,7 +4,7 @@ import Vector2 from "../structs/Vector2.js";
 import Material from "../graphics/Material.js";
 import { player, scene } from "../sceneSetup.js";
 import { game } from "../index.js";
-import Direction from "../search2.js";
+import Direction from "../search.js";
 
 export default class Ghost extends GameObject {
   constructor(position, mesh) {
