+++ index.html
@@ -2,12 +2,14 @@
   <head>
     <script src="libaries/p5.min.js"></script>
     <script src="libaries/quaternion.min.js"></script>
-    <script src="https://cdn.jsdelivr.net/npm/@tensorflow/tfjs/dist/tf.min.js"></script>
     <script src="index.js" type="module"></script>
     <link rel="stylesheet" href="style.css" />
     <title>Best game</title>
   </head>
   <body>
+    <!-- <nav>
+      <h1 class="game-title">Trophy Haunt</h1>
+    </nav> -->
     <main>
       <div style="display: flex; flex-direction: row">
         <div id="canvas"></div>
@@ -33,5 +35,16 @@
         </div>
       </div>
     </main>
+
+    <footer>
+      <a href="https://github.com/Enderpotato/Gamejam2023.git" target="_blank">
+        <img
+          src="assets/imgs/github.png"
+          alt="github"
+          height="40px"
+          width="40px"
+        />
+      </a>
+    </footer>
   </body>
 </html>
