+++ style.css
@@ -1,5 +1,37 @@
 html {
   background-color: rgb(100, 100, 100);
+  padding: 0;
+  margin: 0;
+}
+
+canvas {
+  display: block;
+  margin: 0;
+}
+
+nav {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 1rem;
+  background-color: wheat;
+}
+
+footer {
+  background-color: aquamarine;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 1rem, 0, 1rem, 0;
+  width: 100%;
+  position: fixed;
+  bottom: 0;
+}
+
+.game-title {
+  font-size: 2rem;
+  font-weight: bold;
+  margin: 0;
 }
 
 #camera-debug,
@@ -13,6 +45,7 @@ html {
 
 #debug-screen {
   display: none;
+  width: 100%;
   padding: 1rem;
   background-color: skyblue;
   /* Add other styles as needed */
